Minor Edit to Assest List
Joy did a fantastic job with the assest list, very thankful for their hard week. I know we can get way more done, but these are the bare essentials for the game. In addition, life happens. This game isn't a dedicated project, it's for a class. Of course, we're going to elevate the concept and make it awesome, but it's not a guaranteed time commitment. I want to focus on finished and functional before worrying about putting out the next big hit.