Do a barrel roll!
I implemented the character being able to do a dodge roll. For now, the roll doesn't have any animations or make the player invincible, but it shows how it would work in game for the prototype. I'm going to try to get the player to be invincible while doing a dodge roll and begin working on the burrow mechanic. See ya next update!