Invincibility Added(DodgeRoll)
Added invincibilty for the dodge roll and more force to the dodge roll itself. I need to continue working on the burrow mechanic, but there's some sound logic there. I also need to make it so I can interact when a burrow is created. Yet, there's a huge amount of progress and it's only been worked on for about 10 hours. crazyyyyyyyy