BurrowParticle System Bug Fixed
For some reason the player was not having the particle system play when burrowing, turns out I just needed to adjust the hierarchy. I do believe I was also able to get the trail rendering working because of it, but it looked like an elder demon, so I canned it. Next time, will be getting bullets to go through the burrows.