Reload Update!
Made some changes to the reload functionality. Previously you were only able to reload when you were out of bullets. That's not realistic. So players can now reload as long as they are under the max ammo capacity.

I also added the UI that gives a few different prompts based on the players ammo count. If the player has no bullets in the chamber but has ammo, the player will receive a message telling them to reload. Likewise, if the player has no bullets in the chamber and has no ammo, you get a prompt saying you need to find bullets.

If the player presses reload, a new message stating that the player is reloading pops up. Overall a badly needed update. Next is getting the NPC's dialogue to function properly. Till next time traveller